COM-3302 - refresh potentialTrainees onMounted profilOrganization component + onrefresh
