+++ src/core/components/courses/ProfileOrganization.vue
@@ -34,8 +34,8 @@
       @update="updateCourse('estimatedStartDate')" />
     <ni-company-table v-if="isCourseInter && isVendorInterface" :course="course" @refresh="refreshCourse"
       :can-edit="isAdmin" :loading="courseLoading" />
-    <ni-trainee-table :can-edit="canEditTrainees" :loading="courseLoading" @refresh="refreshCourse"
-      @update="updateCourse('maxTrainees')" :validations="v$.course" />
+    <ni-trainee-table :can-edit="canEditTrainees" :loading="courseLoading" :potential-trainees="potentialTrainees"
+      @refresh="refreshCourse" @update="updateCourse('maxTrainees')" :validations="v$.course" />
     <q-page-sticky expand position="right">
       <course-history-feed v-if="displayHistory" @toggle-history="toggleHistory" :course-histories="courseHistories"
         @load="updateCourseHistories" ref="courseHistoryFeed" />
@@ -111,7 +111,7 @@
 
 <script>
 import { useStore } from 'vuex';
-import { computed, ref, toRefs, watch } from 'vue';
+import { computed, onMounted, ref, toRefs, watch } from 'vue';
 import { copyToClipboard } from 'quasar';
 import useVuelidate from '@vuelidate/core';
 import { required } from '@vuelidate/validators';
@@ -221,6 +221,7 @@ export default {
     const SALES_REPRESENTATIVE = ref('salesRepresentative');
     const COMPANY_REPRESENTATIVE = ref('companyRepresentative');
     const courseHistoryFeed = ref(null);
+    const potentialTrainees = ref([]);
 
     const course = computed(() => $store.state.course.course);
 
@@ -237,6 +238,25 @@ export default {
       followUpMissingInfo,
       downloadAttendanceSheet,
     } = useCourses(course);
+
+    const getPotentialTrainees = async () => {
+      try {
+        let query;
+
+        if (isClientInterface) query = { companies: get(loggedUser.value, 'company._id') };
+        else query = { companies: course.value.companies.map(c => c._id) };
+
+        const queryCompaniesIsNotEmpty = Array.isArray(query.companies) ? !!query.companies.length : !!query.companies;
+        potentialTrainees.value = queryCompaniesIsNotEmpty ? Object.freeze(await Users.learnerList(query)) : [];
+      } catch (error) {
+        potentialTrainees.value = [];
+        console.error(error);
+      }
+    };
+
+    onMounted(async () => {
+      await getPotentialTrainees();
+    });
 
     const loggedUser = computed(() => $store.state.main.loggedUser);
 
@@ -372,6 +392,7 @@ export default {
           await getCourseHistories();
           courseHistoryFeed.value.resumeScroll();
         }
+        await getPotentialTrainees();
       } catch (e) {
         console.error(e);
       } finally {
@@ -709,6 +730,7 @@ export default {
       // Data
       INTRA,
       trainerOptions,
+      potentialTrainees,
       salesRepresentativeOptions,
       companyRepresentativeOptions,
       courseLoading,
